Remove settings.xml from travis config
Usually we'd use the settings.xml from the parent project but it seems we already
have one in this project with bucketeer specific variables. We should probably
remove it and let the parent project's be the one that's used, but I'm pushing
that off for now since it's not related to the work in hand. The one with bucketeer
variables is probably referenced somewhere in the documentation as an example.